Correct celery configuration and queue issues
