+++ banzai/celeryconfig.py
@@ -1,0 +1,3 @@
+broker_url = 'redis://redis:6379/0'
+imports = ('banzai.main', 'banzai.celery',)
+worker_prefetch_multiplier = 1
