+++ docker-compose.yml
@@ -51,6 +51,7 @@ services:
     image: ${DOCKER_IMG}
     container_name: banzai-celery-workers
     entrypoint: ["celery", "-A", "banzai", "worker",
+                 "--hostname", "banzai-celery-worker",
                  "-l", "info"]
     mem_limit: '8g'
     depends_on:
@@ -64,7 +65,7 @@ services:
       - banzai-e2e
     environment:
       RETRY_DELAY: "0"
-      REDIS_HOST: "redis"
+      REDIS_HOST: "redis://redis:6379/0"
     labels:
       io.rancher.container.pull_image: always
       io.rancher.sidekicks: banzai-e2e
@@ -89,7 +90,7 @@ services:
                  "--broker-url=broker"]
     environment:
       DB_ADDRESS: sqlite:////archive/engineering/test.db
-      REDIS_HOST: "redis"
+      REDIS_HOST: "redis://redis:6379/0"
       FITS_BROKER_URL: "broker"
     volumes_from:
     - banzai-e2e
